moved fakeData array into its own component file. Also moved rooms.js into config folder
