+++ client/src/App.js
@@ -2,43 +2,9 @@ import React, { Component } from "react";
 import "./App.css";
 import axios from "axios";
 import { Button, Modal } from "react-materialize";
-import SendMessageBtn from "./Components/sendMessageBtn";
+import SendMessageBtn from "./Components/SendMsgBtn/sendMessageBtn";
+import fakeData from "./Components/FakeData/fakeData";
 
-//Below we create an object that contains a conversation that is used to display in the view when the user toggles the conversation. A developer may add or replace individual messages to render on the landing page. 
-const fakeData = [
-  {
-    _id: 1,
-    name: "Sven",
-    message: "User result came in. They were decent."
-  },
-  {
-    _id: 2,
-    name: "Charles",
-    message: "Ha! nice!"
-  },
-  {
-    _id: 3,
-    name: "Rob",
-    message:
-      "Great to hear. We beat the competitors, but we also blew the budget. Oops."
-  },
-  {
-    _id: 4,
-    name: "Carlita",
-    message:
-      "Our safisfaction survey results just came in too. They aren't...terrible."
-  },
-  {
-    _id: 5,
-    name: "Juan",
-    message: "Let's try Rob's suggestion..."
-  },
-  {
-    _id: 6,
-    name: "Carol",
-    message: "I like that. Are there other ideas to consider?"
-  }
-];
 
 //Here we employ React's component.state property (by way of state.isReal) to track whethere the conversation that is displayed is the real or decoy converstaion.
 class App extends Component {
