+++ client/src/Components/FakeData/fakeData.js
@@ -1,0 +1,38 @@
+//Below we create an object that contains a conversation that is used to display in the view when the user toggles the conversation. A developer may add or replace individual messages to render on the landing page. 
+
+const fakeData = [
+  {
+    _id: 1,
+    name: "Sven",
+    message: "User result came in. They were decent."
+  },
+  {
+    _id: 2,
+    name: "Charles",
+    message: "Ha! nice!"
+  },
+  {
+    _id: 3,
+    name: "Rob",
+    message:
+      "Great to hear. We beat the competitors, but we also blew the budget. Oops."
+  },
+  {
+    _id: 4,
+    name: "Carlita",
+    message:
+      "Our safisfaction survey results just came in too. They aren't...terrible."
+  },
+  {
+    _id: 5,
+    name: "Juan",
+    message: "Let's try Rob's suggestion..."
+  },
+  {
+    _id: 6,
+    name: "Carol",
+    message: "I like that. Are there other ideas to consider?"
+  }
+];
+
+export default fakeData;
