+++ config/rooms.js
